a lot of changes don't want to specify

+++ pegasusX/apps/retailer-app-android/app/src/main/java/com/pegasusx/retailer/ui/screens/product/ProductDetailScreen.kt
@@ -127,7 +127,7 @@ fun ProductDetailScreen(
                                 pack = "Single",
                                 packCount = 1,
                                 weightPerUnit = "",
-                                price = (product.price ?: 0).toDouble(),
+                                price = (product.price ?: 0).toLong(),
                             )
                             onAddToCart(product, variant)
                             scope.launch { snackbarHostState.showSnackbar("${product.name} added to cart") }
